requirements.txt
streamlit
pandas
numpy
scikit-learn
seaborn
matplotlib
google-generativeai
joblib